fix(pixels): fix empty brightness define and blit types
Do not register tw_max_brightness as a Soong config variable unless TW_MAX_BRIGHTNESS is non-empty. This prevents Soong from emitting -DTW_MAX_BRIGHTNESS=, allowing TWRP to use runtime sysfs discovery without producing an invalid macro definition.

Cast GRSurface dimensions to int before std::min compares them with signed source coordinates. Keep both fixes in the modified snapshots and their generated unified patches.

+++ patches/files/modified/bootable/recovery/minuitwrp/graphics.cpp
@@ -326,8 +326,8 @@ void gr_blit(gr_surface source, int sx, int sy, int w, int h, int dx, int dy)
     }
     const int requested_w = w;
     const int requested_h = h;
-    w = std::min(w, surface->width - sx);
-    h = std::min(h, surface->height - sy);
+    w = std::min(w, static_cast<int>(surface->width) - sx);
+    h = std::min(h, static_cast<int>(surface->height) - sy);
     dx += (requested_w - w) / 2;
     dy += (requested_h - h) / 2;
